chore: move modal component into its own component
